[BUGFIX] Fix order of FE initialization in eid linkhandlerPreview
The linkhandlerPreview script had an error that resultet in TCA not being
loaded before the rootline was calculated. This means that relation fields
like the media field on pages (which is a sys_file_reference relation) was
not correctly resolved.
Since this invalid rootline is then cached, all subsequent requests to that
page will result in an exception when the media field now does not contain
a sys_file_reference, but instead the direct value of the media field (which
will be the number of relatiosn due to how TCA works).